requirements.txt
streamlit
pymupdf4llm
langchain
langchain-community
langchain-text-splitters
openai
python-dotenv
pydantic